move strip_bad_sf_chars to util and remove gamepedia.py
